+++ example/android/app/src/main/kotlin/com/letsar/sh/example/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.letsar.sh.example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}